snapshot: interfaces changes, break everything
