+++ level/level0.go
@@ -7,6 +7,9 @@ import (
 	"fmt"
 	iiifcompliance "github.com/go-iiif/go-iiif/v4/compliance"
 	iiifconfig "github.com/go-iiif/go-iiif/v4/config"
+	iiifimage "github.com/go-iiif/go-iiif/v4/image"
+	iiifprofile "github.com/go-iiif/go-iiif/v4/profile"
+	iiifservice "github.com/go-iiif/go-iiif/v4/service"
 	_ "log"
 )
 
@@ -65,4 +68,29 @@ func NewLevel0(config *iiifconfig.Config, endpoint string) (*Level0, error) {
 
 func (l *Level0) Compliance() iiifcompliance.Compliance {
 	return l.compliance
+}
+
+func (l *Level0) Profile(endpoint string, image iiifimage.Image) iiifprofile.Profile {
+
+	dims, err := image.Dimensions()
+
+	if err != nil {
+		return nil, fmt.Errorf("Failed to derive dimensions for image, %w", err)
+	}
+
+	p := iiifprofile.Profile{
+		Context:  "http://iiif.io/api/image/2/context.json",
+		Id:       fmt.Sprintf("%s/%s", endpoint, image.Identifier()),
+		Type:     "iiif:Image",
+		Protocol: "http://iiif.io/api/image",
+		Width:    dims.Width(),
+		Height:   dims.Height(),
+		Profile: []interface{}{
+			"http://iiif.io/api/image/2/level0.json",
+			level,
+		},
+		Services: []iiifservice.Service{},
+	}
+
+	return &p, nil
 }
